# specs/openapi-spec.yaml
openapi: 3.0.3
info:
  title: Teams API
  description: API for managing soccer teams, players, trainers, events, and equipment
  version: 1.0.0
  contact:
    name: Teams API Support
    email: support@teams-api.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.teams.com
    description: Production server
# Authentication is optional for most endpoints, but required for some
security: []
paths:
...
  /api/groups/{groupId}:
    get:
      operationId: getGroupById
      summary: Get group by ID
      description: Retrieve a specific group by ID
      tags:
        - Groups
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Group found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '404':
          description: Group not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      operationId: updateGroupById
      summary: Update group by ID
      description: Update a specific group by ID
      tags:
        - Groups
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGroupRequest'
      responses:
        '200':
          description: Group updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Group not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      operationId: deleteGroupById
      summary: Delete group by ID
      description: Delete a specific group by ID
      tags:
        - Groups
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Group deleted successfully
        '404':
          description: Group not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/groups/{groupId}/playing-modes:
    x-aliases:
      - /api/squads/{squadId}/playing-modes
    get:
      operationId: getGroupPlayingModes
      summary: Get playing modes for a group
      description: >-
        Retrieve all Playing Modes belonging to a specific group. This is a
        group-scoped subresource; there is no top-level playing-modes
        collection.
      tags:
        - Groups
      security:
        - bearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved playing modes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PlayingMode'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not authorized to access this group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Group not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      operationId: createGroupPlayingMode
      summary: Create playing mode for a group
      description: >-
        Add a new Playing Mode to a group. Requires admin or trainer role in
        the group (same authorization as event/team modification). The first
        Playing Mode created for a group is automatically marked default.
      tags:
        - Groups
      security:
        - bearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePlayingModeRequest'
      responses:
        '201':
          description: Playing mode created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlayingMode'
        '400':
          description: >-
            Invalid request data: name is empty, numberOfPeriods or
            periodLengthMinutes is not an integer >= 1, or
            minimumPeriodsPerPlayer is not an integer between 0 and
            numberOfPeriods.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: >-
            Not authorized: requires admin or trainer role in this group.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Group not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login or /auth/register
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    Group:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        club:
          type: string
          nullable: true
        category:
          $ref: '#/components/schemas/GroupCategory'
        description:
          type: string
          nullable: true
        trainers:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              email:
                type: string
                format: email
        members:
          type: array
          items:
            oneOf:
              - $ref: '#/components/schemas/Player'
              - $ref: '#/components/schemas/Trainer'
        periods:
          type: array
          items:
            $ref: '#/components/schemas/Period'
        matchPlanningEnabled:
          type: boolean
          default: false
          description: Opt-in flag for Playing Modes/Formations/lineup planning.
        playingModes:
          type: array
          items:
            $ref: '#/components/schemas/PlayingMode'
          description: Group-scoped only; never shared with other groups.
        matchFormats:
          type: array
          items:
            $ref: '#/components/schemas/PlayingMode'
          description: >-
            Squad alias of playingModes; returned on /api/squads routes with
            the same values. Stored as playingModes.
        formations:
          type: array
          items:
            $ref: '#/components/schemas/Formation'
          description: Group-scoped only; never shared with other groups.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
...
    Period:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
      required:
        - id
        - name
        - startDate
        - endDate
...
    UpdateGroupRequest:
      type: object
      properties:
        name:
          type: string
        club:
          type: string
          nullable: true
        category:
          $ref: '#/components/schemas/GroupCategory'
        description:
          type: string
          nullable: true
        trainers:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              email:
                type: string
                format: email
        matchPlanningEnabled:
          type: boolean
          description: >-
            Toggling this does not delete existing Playing Modes, Formations,
            or Team lineup data. Requires admin or trainer role.
    GroupCategory:
      type: string
      enum: [A, B, C, D9, D7, E, F, G, FF9, FF11, FF14, FF17]
      description: Swiss Football Association (SFV) team category.
...
    PlayingMode:
      type: object
      description: >-
        Group-scoped subresource (see Group.playingModes). Never referenced
        across groups.
      properties:
        id:
          type: string
        name:
          type: string
          example: "4x20"
        numberOfPeriods:
          type: integer
          minimum: 1
          example: 4
        periodLengthMinutes:
          type: integer
          minimum: 1
          example: 20
        minimumPeriodsPerPlayer:
          type: integer
          minimum: 0
          example: 2
          default: 0
          description: >-
            Minimum number of distinct periods in which every selected player
            should have a field assignment. Must not exceed numberOfPeriods.
            Zero disables the minimum-playing-time rule. Violations are
            advisory and do not prevent saving a team lineup.
        isDefault:
          type: boolean
          default: false
          description: Exactly one Playing Mode per group may be true.
        playersOnField:
          type: integer
          minimum: 1
        minPlayersPerTeam:
          type: integer
          minimum: 1
        maxPlayersPerTeam:
          type: integer
          minimum: 1
        origin:
          type: string
      required:
        - id
        - name
        - numberOfPeriods
        - periodLengthMinutes
        - minimumPeriodsPerPlayer
    CreatePlayingModeRequest:
      type: object
      properties:
        name:
          type: string
        numberOfPeriods:
          type: integer
          minimum: 1
        periodLengthMinutes:
          type: integer
          minimum: 1
        minimumPeriodsPerPlayer:
          type: integer
          minimum: 0
          default: 0
          description: >-
            Must not exceed numberOfPeriods. Zero disables the rule.
        playersOnField:
          type: integer
          minimum: 1
        minPlayersPerTeam:
          type: integer
          minimum: 1
        maxPlayersPerTeam:
          type: integer
          minimum: 1
        origin:
          type: string
      required:
        - name
        - numberOfPeriods
        - periodLengthMinutes
        - minimumPeriodsPerPlayer
...
    PositionCode:
      type: string
      description: >-
        Language-neutral position abbreviation (same catalog as used in
        games like FIFA). Translating a code into a localized label (e.g.
        English "LB" for "left back" vs. German "LV" for "linker
        Verteidiger") is a frontend i18n concern; the backend only stores
        and validates the code.
      enum:
        - GK
        - LB
        - CB
        - RB
        - LWB
        - RWB
        - CDM
        - CM
        - CAM
        - LM
        - RM
        - LW
        - RW
        - CF
        - ST
    FormationSlot:
      type: object
      properties:
        id:
          type: string
        positionCode:
          $ref: '#/components/schemas/PositionCode'
      required:
        - id
        - positionCode
    Formation:
      type: object
      description: >-
        Group-scoped subresource (see Group.formations). Never referenced
        across groups.
      properties:
        id:
          type: string
        name:
          type: string
          example: "3-3"
        slots:
          type: array
          minItems: 2
          items:
            $ref: '#/components/schemas/FormationSlot'
          description: >-
            Exactly one slot must have positionCode GK; other codes may
            repeat (e.g. two CB slots), distinguished only by array order.
      required:
        - id
        - name
        - slots
...
    Player:
      type: object
      properties:
        id:
          type: string
        roles:
          type: array
          items:
            $ref: '#/components/schemas/GroupRole'
        firstName:
          type: string
        lastName:
          type: string
        birthDate:
          type: string
          format: date
          description: Canonical player birth date; birthYear is derived from this value.
        birthYear:
          type: integer
          readOnly: true
          deprecated: true
          description: Derived from birthDate. Use birthDate for writes.
        level:
          type: integer
          minimum: 1
          maximum: 5
        preferredShirtNumber:
          type: integer
          minimum: 1
        status:
          type: string
          enum: [active, trial, inactive]
          default: active
        guardians:
          type: array
          items:
            $ref: '#/components/schemas/Guardian'
      required:
        - id
        - roles
        - firstName
        - lastName
        - level
...
    Trainer:
      type: object
      properties:
        id:
          type: string
        roles:
          type: array
          items:
            $ref: '#/components/schemas/GroupRole'
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
          nullable: true
      required:
        - id
        - roles
        - firstName
        - lastName
    Guardian:
      type: object
      properties:
        id:
          type: string
        roles:
          type: array
          items:
            $ref: '#/components/schemas/GroupRole'
        groupId:
          type: string
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        email:
          type: string
          format: email
          nullable: true
        userId:
          type: string
          deprecated: true
          description: Legacy guardian account identifier; use the guardian member id.
      required:
        - id
        - groupId
        - roles
    GroupRole:
      type: string
      enum: [admin, trainer, guardian, player]
...